requirements.txt
requests
pandas
python-dotenv
apache-airflow
psycopg2-binary